change alert file structure & improve animation
