+++ library/src/lib/alert/alert.module.ts
@@ -4,8 +4,8 @@ import { IconModule } from '../icon/icon.module';
 import { UtilsModule } from '../utils/utils.module';
 
 import { AlertComponent } from './alert.component';
-import { AlertService } from './alert.service';
-import { AlertContainerComponent } from './alert-container.component';
+import { AlertService } from './alert-service/alert.service';
+import { AlertContainerComponent } from './alert-utils/alert-container.component';
 
 @NgModule({
     declarations: [AlertComponent, AlertContainerComponent],
