+++ library/src/public_api.ts
@@ -8,10 +8,10 @@ export * from './lib/fundamental-ngx.module';
 
 export * from './lib/action-bar/action-bar.module';
 export * from './lib/alert/alert.module';
-export * from './lib/alert/alert.service';
+export * from './lib/alert/alert-service/alert.service';
 export * from './lib/alert/alert.component';
-export * from './lib/alert/alert-config';
-export * from './lib/alert/alert-ref';
+export * from './lib/alert/alert-utils/alert-config';
+export * from './lib/alert/alert-utils/alert-ref';
 export * from './lib/badge-label/badge-label.module';
 export * from './lib/breadcrumb/breadcrumb.module';
 export * from './lib/button/button.module';
